many changes to make it train fast on cpu  machine
mainly I removed the large number of experiences from the beginning and instead start with the same as batch size. I made the CNN more efficient. I decreased the batch size to 4. do not know the best batch size it is a trade off between time spent training and time spent getting experience, and a faster machine might benefit with a larger batchsize.